pyproject.toml
[project]
name = "prd-code-verifier"
version = "0.1.0"
description = "PRD Code Verifier - A web application to verify code against documentation using AI"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "gowrav-vishwakarma", email = "gowrav.vishwakarma@gmail.com"}
]
license = {text = "MIT"}
repository = "https://github.com/gowrav-vishwakarma/prd-code-verifier"
homepage = "https://github.com/gowrav-vishwakarma/prd-code-verifier"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.9.0",
    "openai>=1.3.0",
    "google-generativeai>=0.3.0",
    "ollama>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
